requirements.txt
langchain==0.3.14
langchain-openai==0.2.14
langchain-community==0.3.14
langchain-chroma==0.1.4
openai==1.59.3
httpx==0.27.0
python-dotenv==1.0.1
gradio==5.9.1
pypdf==5.1.0